requirements.txt
Flask==3.1.0
pymongo==4.10.1
python-dotenv==1.0.1
requests==2.32.3
